Fix whitespace and line length.
Part of this changes the call of defaultData() in "__main__"
to generically pass its arguments as *args
to the main checkAstrometry.run() without explicitly listing
each of the arguments on the LHS and then RHS a line later.
This was nominally motivated by the line length being too long
but seems just like better practice anyway.