Calculate sph dist using Haversine instead of arccos.
When using arccos to calculate distances on a sphere, points with
very close (or identical) coordinates led to round-off errors which
led to taking arccos(x>1.0).
Using haversine is the better choice and resolves this problem.
Vincenty is unlikely to be needed for this application, as calculating
distances across to opposites of the sphere is not part of the current
scope.