Define N_CORES in runHsc script to control number cores.
The default was '--cores 16' which was too much memory for some systems.
At an optimistic 1 GB/core, and a conservative 2 GB/core, the N_CORES
is now set at 4, which should run comfortably on all modern machines.
For faster processing on a larger machine, one could set this up to 16.