Correct units in plotPhotErrModelFit label.
The manually-specified "\mathrm{{[mmag]}}" unit text was being overwritten
(or pushed aside?) by the automatic astropy.units-generated unit text.

We expect sigmaSys phot to potentially be on the order of ~10 mmag.
Thus we explicitly extract 1000*to(u.mag) and manually format mmag
to get the desired display.

None of this affects the actual value stored in the JSON file which
was and remains correct.