Add import matplotlib; matplotlib.use('Agg') to bin.src/validateDrp.py
We need to be able to run on displayless nodes, even if they
have misconfigured with a default X-enabled backend.
We set the backend explicitly in this command-line script.
This means that we're set for the invocation of this script
but because it's not in the modules, people can still import
all of the modules of validate_drp without triggering any problems
with prematurely setting a matplotlib backend.

The issue is that matplotlib backends can only be set once per Python session.